+++ tests/components/plugwise/test_climate.py
@@ -1,6 +1,7 @@
 """Tests for the Plugwise Climate integration."""
 
 from datetime import timedelta
+import math
 from unittest.mock import MagicMock, patch
 
 from plugwise.exceptions import PlugwiseError
@@ -13,6 +14,9 @@ from homeassistant.util.dt import utcnow
 
 from tests.common import MockConfigEntry, async_fire_time_changed
 
+HA_PLUGWISE_SMILE_ASYNC_UPDATE = (
+    "homeassistant.components.plugwise.coordinator.Smile.async_update"
+)
 TEST_HOST = "1.1.1.1"
 TEST_PASSWORD = "test_password"
 
@@ -24,7 +28,10 @@ async def test_adam_climate_entity_attributes(
     state = hass.states.get("climate.zone_lisa_wk")
     assert state
     assert state.state == HVACMode.AUTO
-    assert state.attributes["hvac_modes"] == [HVACMode.HEAT, HVACMode.AUTO]
+    assert state.attributes["hvac_modes"] == [
+        HVACMode.AUTO,
+        HVACMode.HEAT,
+    ]
     # hvac_action is not asserted as the fixture is not in line with recent firmware functionality
 
     assert "preset_modes" in state.attributes
@@ -37,12 +44,17 @@ async def test_adam_climate_entity_attributes(
     assert state.attributes["temperature"] == 21.5
     assert state.attributes["min_temp"] == 0.0
     assert state.attributes["max_temp"] == 99.9
-    assert state.attributes["target_temp_step"] == 0.5
+    assert math.isclose(
+        state.attributes["target_temp_step"], 0.1, rel_tol=1e-09, abs_tol=1e-09
+    )
 
     state = hass.states.get("climate.zone_thermostat_jessie")
     assert state
     assert state.state == HVACMode.AUTO
-    assert state.attributes["hvac_modes"] == [HVACMode.HEAT, HVACMode.AUTO]
+    assert state.attributes["hvac_modes"] == [
+        HVACMode.AUTO,
+        HVACMode.HEAT,
+    ]
     # hvac_action is not asserted as the fixture is not in line with recent firmware functionality
 
     assert "preset_modes" in state.attributes
@@ -54,7 +66,9 @@ async def test_adam_climate_entity_attributes(
     assert state.attributes["temperature"] == 15.0
     assert state.attributes["min_temp"] == 0.0
     assert state.attributes["max_temp"] == 99.9
-    assert state.attributes["target_temp_step"] == 0.5
+    assert math.isclose(
+        state.attributes["target_temp_step"], 0.1, rel_tol=1e-09, abs_tol=1e-09
+    )
 
 
 async def test_adam_2_climate_entity_attributes(
@@ -65,13 +79,21 @@ async def test_adam_2_climate_entity_attributes(
     assert state
     assert state.state == HVACMode.HEAT
     assert state.attributes["hvac_action"] == "heating"
-    assert state.attributes["hvac_modes"] == [HVACMode.HEAT, HVACMode.AUTO]
+    assert state.attributes["hvac_modes"] == [
+        HVACMode.OFF,
+        HVACMode.AUTO,
+        HVACMode.HEAT,
+    ]
 
     state = hass.states.get("climate.lisa_badkamer")
     assert state
     assert state.state == HVACMode.AUTO
     assert state.attributes["hvac_action"] == "heating"
-    assert state.attributes["hvac_modes"] == [HVACMode.HEAT, HVACMode.AUTO]
+    assert state.attributes["hvac_modes"] == [
+        HVACMode.OFF,
+        HVACMode.AUTO,
+        HVACMode.HEAT,
+    ]
 
 
 async def test_adam_3_climate_entity_attributes(
@@ -79,13 +101,59 @@ async def test_adam_3_climate_entity_attributes(
 ) -> None:
     """Test creation of adam climate device environment."""
     state = hass.states.get("climate.anna")
-
     assert state
-    assert state.state == HVACMode.HEAT_COOL
+    assert state.state == HVACMode.COOL
     assert state.attributes["hvac_action"] == "cooling"
     assert state.attributes["hvac_modes"] == [
-        HVACMode.HEAT_COOL,
+        HVACMode.OFF,
         HVACMode.AUTO,
+        HVACMode.COOL,
+    ]
+    data = mock_smile_adam_3.async_update.return_value
+    data.devices["da224107914542988a88561b4452b0f6"][
+        "select_regulation_mode"
+    ] = "heating"
+    data.devices["ad4838d7d35c4d6ea796ee12ae5aedf8"]["mode"] = "heat"
+    data.devices["056ee145a816487eaa69243c3280f8bf"]["binary_sensors"][
+        "cooling_state"
+    ] = False
+    data.devices["056ee145a816487eaa69243c3280f8bf"]["binary_sensors"][
+        "heating_state"
+    ] = True
+    with patch(HA_PLUGWISE_SMILE_ASYNC_UPDATE, return_value=data):
+        async_fire_time_changed(hass, utcnow() + timedelta(minutes=1))
+        await hass.async_block_till_done()
+    state = hass.states.get("climate.anna")
+    assert state
+    assert state.state == HVACMode.HEAT
+    assert state.attributes["hvac_action"] == "heating"
+    assert state.attributes["hvac_modes"] == [
+        HVACMode.OFF,
+        HVACMode.AUTO,
+        HVACMode.HEAT,
+    ]
+    data = mock_smile_adam_3.async_update.return_value
+    data.devices["da224107914542988a88561b4452b0f6"][
+        "select_regulation_mode"
+    ] = "cooling"
+    data.devices["ad4838d7d35c4d6ea796ee12ae5aedf8"]["mode"] = "cool"
+    data.devices["056ee145a816487eaa69243c3280f8bf"]["binary_sensors"][
+        "cooling_state"
+    ] = True
+    data.devices["056ee145a816487eaa69243c3280f8bf"]["binary_sensors"][
+        "heating_state"
+    ] = False
+    with patch(HA_PLUGWISE_SMILE_ASYNC_UPDATE, return_value=data):
+        async_fire_time_changed(hass, utcnow() + timedelta(minutes=1))
+        await hass.async_block_till_done()
+    state = hass.states.get("climate.anna")
+    assert state
+    assert state.state == HVACMode.COOL
+    assert state.attributes["hvac_action"] == "cooling"
+    assert state.attributes["hvac_modes"] == [
+        HVACMode.OFF,
+        HVACMode.AUTO,
+        HVACMode.COOL,
     ]
 
 
@@ -176,6 +244,60 @@ async def test_adam_climate_entity_climate_changes(
         )
 
 
+async def test_adam_climate_off_mode_change(
+    hass: HomeAssistant,
+    mock_smile_adam_4: MagicMock,
+    init_integration: MockConfigEntry,
+) -> None:
+    """Test handling of user requests in adam climate device environment."""
+    state = hass.states.get("climate.slaapkamer")
+    assert state
+    assert state.state == HVACMode.OFF
+    await hass.services.async_call(
+        "climate",
+        "set_hvac_mode",
+        {
+            "entity_id": "climate.slaapkamer",
+            "hvac_mode": "heat",
+        },
+        blocking=True,
+    )
+    assert mock_smile_adam_4.set_schedule_state.call_count == 1
+    assert mock_smile_adam_4.set_regulation_mode.call_count == 1
+    mock_smile_adam_4.set_regulation_mode.assert_called_with("heating")
+
+    state = hass.states.get("climate.kinderkamer")
+    assert state
+    assert state.state == HVACMode.HEAT
+    await hass.services.async_call(
+        "climate",
+        "set_hvac_mode",
+        {
+            "entity_id": "climate.kinderkamer",
+            "hvac_mode": "off",
+        },
+        blocking=True,
+    )
+    assert mock_smile_adam_4.set_schedule_state.call_count == 1
+    assert mock_smile_adam_4.set_regulation_mode.call_count == 2
+    mock_smile_adam_4.set_regulation_mode.assert_called_with("off")
+
+    state = hass.states.get("climate.logeerkamer")
+    assert state
+    assert state.state == HVACMode.HEAT
+    await hass.services.async_call(
+        "climate",
+        "set_hvac_mode",
+        {
+            "entity_id": "climate.logeerkamer",
+            "hvac_mode": "heat",
+        },
+        blocking=True,
+    )
+    assert mock_smile_adam_4.set_schedule_state.call_count == 1
+    assert mock_smile_adam_4.set_regulation_mode.call_count == 2
+
+
 async def test_anna_climate_entity_attributes(
     hass: HomeAssistant, mock_smile_anna: MagicMock, init_integration: MagicMock
 ) -> None:
@@ -185,8 +307,8 @@ async def test_anna_climate_entity_attributes(
     assert state.state == HVACMode.AUTO
     assert state.attributes["hvac_action"] == "heating"
     assert state.attributes["hvac_modes"] == [
-        HVACMode.HEAT,
         HVACMode.AUTO,
+        HVACMode.HEAT,
     ]
 
     assert "no_frost" in state.attributes["preset_modes"]
@@ -198,7 +320,9 @@ async def test_anna_climate_entity_attributes(
     assert state.attributes["temperature"] == 20.5
     assert state.attributes["min_temp"] == 4.0
     assert state.attributes["max_temp"] == 30.0
-    assert state.attributes["target_temp_step"] == 0.5
+    assert math.isclose(
+        state.attributes["target_temp_step"], 0.1, rel_tol=1e-09, abs_tol=1e-09
+    )
 
 
 async def test_anna_2_climate_entity_attributes(
@@ -212,8 +336,8 @@ async def test_anna_2_climate_entity_attributes(
     assert state.state == HVACMode.AUTO
     assert state.attributes["hvac_action"] == "cooling"
     assert state.attributes["hvac_modes"] == [
-        HVACMode.HEAT_COOL,
         HVACMode.AUTO,
+        HVACMode.HEAT_COOL,
     ]
     assert state.attributes["supported_features"] == 18
     assert state.attributes["target_temp_high"] == 24.0
@@ -231,8 +355,8 @@ async def test_anna_3_climate_entity_attributes(
     assert state.state == HVACMode.AUTO
     assert state.attributes["hvac_action"] == "idle"
     assert state.attributes["hvac_modes"] == [
-        HVACMode.HEAT_COOL,
         HVACMode.AUTO,
+        HVACMode.HEAT_COOL,
     ]
 
 
@@ -271,10 +395,8 @@ async def test_anna_climate_entity_climate_changes(
         {"entity_id": "climate.anna", "hvac_mode": "auto"},
         blocking=True,
     )
-    assert mock_smile_anna.set_schedule_state.call_count == 1
-    mock_smile_anna.set_schedule_state.assert_called_with(
-        "c784ee9fdab44e1395b8dee7d7a497d5", "on"
-    )
+    assert mock_smile_anna.set_schedule_state.call_count == 0
+    # hvac_mode is already auto so not called.
 
     await hass.services.async_call(
         "climate",
@@ -282,16 +404,13 @@ async def test_anna_climate_entity_climate_changes(
         {"entity_id": "climate.anna", "hvac_mode": "heat"},
         blocking=True,
     )
-    assert mock_smile_anna.set_schedule_state.call_count == 2
+    assert mock_smile_anna.set_schedule_state.call_count == 1
     mock_smile_anna.set_schedule_state.assert_called_with(
         "c784ee9fdab44e1395b8dee7d7a497d5", "off"
     )
     data = mock_smile_anna.async_update.return_value
     data.devices["3cb70739631c4d17a86b8b12e8a5161b"]["available_schedules"] = ["None"]
-    with patch(
-        "homeassistant.components.plugwise.coordinator.Smile.async_update",
-        return_value=data,
-    ):
+    with patch(HA_PLUGWISE_SMILE_ASYNC_UPDATE, return_value=data):
         async_fire_time_changed(hass, utcnow() + timedelta(minutes=1))
         await hass.async_block_till_done()
         state = hass.states.get("climate.anna")